Refactor. Moved python_version from mozilla.py to constant.py. Thus mimipy.py and chromium_based.py are not dependent on mozilla.py any more
